requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-json-logger==2.0.7
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
requests==2.31.0 